Add files via upload
Testing of limit switches with external interrupts and debouncing of 500ms there are 2 external interrupts for the 4 limit switches, one for either left or right and one for either top or bottom as left and right cannot be pressed at the same time, same as top and bottom we will use the direction of the motor along with the homing to tell which limit switch is hit. The other 4 external interrupts can be used for the 2 motor encoders which require 2 interrput each.